pyproject.toml
[project]
name = "jtf-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "dagster>=1.12.20",
    "dagster-dbt>=0.28.20",
    "dagster-webserver>=1.12.20",
    "dbt>=1.0.0.40.15",
    "dbt-bigquery>=1.11.1",
    "dbt-duckdb>=1.10.1",
    "duckdb>=1.5.1",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.1.1",
    "pandas>=3.0.1",
    "requests>=2.33.0",
]

[tool.dagster]
module_name = "orch.definitions"
